Make memoryhead double

+++ src/main/java/alluxio/client/file/cache/ShadowCache.java
@@ -29,7 +29,7 @@ import java.util.concurrent.atomic.AtomicReferenceArray;
 public class ShadowCache {
   private static final int DEFAULT_MEMORY = 125;
 
-  private final int mMemoryOverhead; // memory overhead in mb
+  private final double mMemoryOverhead; // memory overhead in mb
   private final int mNumBloomFilter;
   private final long mBloomFilterExpectedInsertions;
   // An array of bloom filters, and each capture a segment of window
@@ -52,12 +52,13 @@ public class ShadowCache {
     this(DEFAULT_MEMORY);
   }
 
-  public ShadowCache(int memoryOverhead) {
+  public ShadowCache(double memoryOverhead) {
     mMemoryOverhead = memoryOverhead;
     // long windowMs = conf.getMs(PropertyKey.USER_CLIENT_CACHE_SHADOW_WINDOW);
     mNumBloomFilter = 4;
     // include the 1 extra working set bloom filter
-    long perBloomFilterMemoryOverhead = mMemoryOverhead * Constants.MB * 8 / (mNumBloomFilter + 1);
+    long perBloomFilterMemoryOverhead =
+        (long) (mMemoryOverhead * Constants.MB * 8 / (mNumBloomFilter + 1));
     // assume 3% Guava default false positive ratio
     mBloomFilterExpectedInsertions =
         (long) ((-perBloomFilterMemoryOverhead * Math.log(2) * Math.log(2)) / Math.log(0.03));
